+++ src/api/errors.ts
@@ -6,6 +6,6 @@ export function invalidPkgField(field: string, type: string): TypeError {
   return error(`Invalid package.json "${field}" field. It has to be of type ${type}`);
 }
 
-export function invalidOption(option: string): TypeError {
-  return error(`Invalid "${option}" option`);
+export function invalidOption(option: string, type: string): TypeError {
+  return error(`Invalid "${option}" option. It has to be of type ${type}`);
 }